[Windows DS] use localized device names
With this fix, all localized device names will be used. In the original code, RtAudio
tries to merge input and output devices into a single entry in the device array. This
might fail with localized device names, because the default input and default output
devices have different names. I guess that's the reason the original code overwrote the
default device name in the first place.  In my changed code, separate array entries are
used for input and output devices. This will preserve the relative order of the devices:
the default device will always come first, and all names will preserve the localization
provided by the OS.

From @tucano42

Closes #231.