ALSA: rewrite of device listing logic to be more similar to aplay
Uses the same logic as the "aplay" utility that comes with ALSA to get the device list,
with some additional filtering for undesired plugins like "plughw" and "null".

Note that the list of devices is scanned and open-probed whenever `getDeviceCount()`,
`getDeviceInfo()`, or `openStream()` are called.

The result is a device list more similar to what is seen in popular software like
Audacity, etc.